% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvnovar.R
\docType{data}
\name{cvnovar}
\alias{cvnovar}
\title{ZTF r-band Cataclysmic Variable/Nova}
\format{
A data frame with 65 observations on the following 3 variables:
\describe{
  \item{t}{heliocentric Julian Day - 2400000}
  \item{m}{magnitude}
  \item{merr}{measurement error standard deviations.}
}
}
\usage{
cvnovar
}
\description{
Time series of a cataclysmic variable/nova object observed in the r-band of the ZTF survey and processed by the ALeRCE broker.ZTF Object code: ZTF18aayzpbr
}
\examples{
data(cvnovar)
plot(cvnovar$t,cvnovar$m,type="l",ylab="",xlab="",col="red")
}
\references{
\insertRef{Forster_2021}{iAR}
}
\keyword{datasets}
