% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldlc.R
\name{foldlc}
\alias{foldlc}
\title{Plotting folded light curves}
\usage{
foldlc(file, f1, plot = TRUE)
}
\arguments{
\item{file}{Matrix with the light curve observations. The first column must have the irregular times, the second column must have the brightness magnitudes and the third column must have the measurement errors.}

\item{f1}{Frequency (1/Period) of the light curve.}

\item{plot}{logical; if TRUE, the function returns the plot of folded time series.}
}
\value{
A matrix whose first column has the folded (phased) observational times.
}
\description{
This function plots a time series folded on its period.
}
\examples{
data(clcep)
f1=0.060033386
foldlc(clcep,f1)
}
