% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iAR.R
\name{iAR-package}
\alias{iAR-package}
\title{iAR: Irregularly Observed Autoregressive Models}
\description{
Description: Data sets, functions and scripts with examples to implement autoregressive
models for irregularly observed time series. The models available in this package
are the irregular autoregressive model (Eyheramendy et al.(2018) <doi:10.1093/mnras/sty2487>),
the complex irregular autoregressive model (Elorrieta et al.(2019) <doi:10.1051/0004-6361/201935560>)
and the bivariate irregular autoregressive model (Elorrieta et al.(2021) <doi:10.1093/mnras/stab1216>)
}
\details{
"_PACKAGE"
}
