% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/08_plot.R
\name{plot}
\alias{plot}
\title{Plot Method for unidata Objects}
\arguments{
\item{x}{An object of class `unidata` or `multidata`. This object must contain the time 
series in the `series` slot and the associated time points in the 
`times` slot.}

\item{...}{Additional arguments passed to the `plot.zoo` function for 
customizing the plot (e.g., titles, colors, etc.).}
}
\description{
This method allows visualizing `unidata` and `multidata` objects using the `plot.zoo` 
function from the `zoo` package. It converts the `unidata` (`multidata`) object into 
a `zoo` object and then applies the plotting function to the time series 
contained in the object.
}
